+++ app/actions/tournaments.ts
@@ -135,3 +135,46 @@ export async function patchTournamentData<K extends keyof TournamentDTO>({
 
   return result;
 }
+
+export async function fetchTournamentsPage(params: {}) {
+  const session = await getSessionData();
+
+  /* IF USER IS UNAUTHORIZED REDIRECT TO HOMEPAGE */
+  if (!session.id) return redirect('/');
+
+  const { page } = params;
+
+  const queryCheck = await TournamentsQuerySchema.safeParse({
+    page: page ? +page : page,
+  });
+
+  if (!queryCheck.success) {
+    return console.log('error');
+  }
+
+  const wrapper = new TournamentsWrapper(apiWrapperConfiguration);
+
+  let data = await wrapper.list({
+    page: 1,
+    pageSize: 30,
+    verified: false,
+  });
+
+  return data.result;
+}
+
+export async function fetchTournamentPage(tournamentId: number | string) {
+  const session = await getSessionData();
+
+  /* IF USER IS UNAUTHORIZED REDIRECT TO HOMEPAGE */
+  if (!session.id) return redirect('/');
+
+  const wrapper = new TournamentsWrapper(apiWrapperConfiguration);
+
+  let data = await wrapper.get({
+    id: tournamentId as number,
+    verified: false
+  })
+
+  return data.result;
+}
